Fix standard language re projection-by-type
This:

 * adds projection-by-type to the list of expressions in
   standard/README.md.

 * migrates a beta-normalization rule for field selection which somehow
   got removed from the other field selection rule.

 * uses two blank lines consistently around judgments

 * makes it clear that you can use an empty record type as a selector in
   a project-by-type expression, by using `{ ss… }` notation instead of
   `{ x : T, ss… }`

 * removes the normalization judgment for `t₀.(s₀) ⇥ t₁.(s₁)` that could
   never match

The last point deserves more explanation.  The expression `t.(s)` can
only typecheck if `s ⇥ { ss… }`, which means that we can always
normalize a `t.(s)` expression to a `t.{ ss… }` expression using the
rule above.  Because the `t₀.(s₀) ⇥ t₁.(s₁)` judgment only matches "If
no other rule matches", and because the other rule always matches, it is
redundant and can be removed.