Fix and improve projection-by-type type inference
This fixes and improves type inference of `t.(s)` expressions.

The bug to be fixed is that previously, you needed the following
preconditions for the rule to match:

    Γ ⊢ s : T₀
    s ⇥ { x : T₀, ss… }

I don't think it's possible to satisfy both rules simultaneously.  I
think it was a mistake for the same symbol T₀ to be used in both rules;
the fix is to be more careful with use of variable names.

The improvement is that, after this change, your selector record doesn't
have to exactly match the types of the fields it's projecting out; it
only has to be judgmentally equal.  I've added a test to capture this
case.