Implement importing paths as Location (#585)
This standardizes native support for reading import locations into Dhall,
as specified by @Gabriel439 in 
https://github.com/dhall-lang/dhall-lang/issues/71#issuecomment-500179115

So e.g. if you do

```dhall
./some/import as Location
```

..then you'd have it read in Dhall as

```dhall
< Local : Text | Remote : Text | Environment : Text | Missing >.Local "./some/import"
```

The main reason why this is useful as a builtin is that we can exploit the
canonicalization and chaining that Dhall uses for imports so that relative
paths stay correct over relative importing of other Dhall expressions.
This sounds like a mouthful, so here's an example (this also includes a test
about it):

- if `./nested/import.dhall` contains `missing`
- if `./nested/importLocation.dhall` contains `./import.dhall as Location`
- and `here.dhall` contains `./nested/importLocation.dhall`
  (note that this is a normal import)
- then resolving `here.dhall` in the current directory will give:
```dhall
< Local : Text | Remote : Text | Environment : Text | Missing >.Local "./nested/import.dhall"

-- Note: `./nested/import.dhall` is never resolved!
```

I also noted that `missing` will _not_ fail to resolve when imported
`as Location`, so e.g. the `?` operator will not have the usual behaviour
(there's a test about this too)
```dhall
(missing as Location) ? 42

-- This resolves to:
< Local : Text | Remote : Text | Environment : Text | Missing >.Missing
```