Treat "multi-lets" as syntactic sugar
Removes any mention of "multi-lets" from the semantics, while adding a
note to the grammar that `let... let... in...` is to be treated as
syntactic sugar for `let... in let... in...`. The resulting standard is
equivalent to the old one, only arguably simpler and easier to follow.

Addresses https://github.com/dhall-lang/dhall-lang/issues/615; see
https://github.com/dhall-lang/dhall-lang/commit/022105eefee605f6e5b862b9d7816862b9db73e2
for the commit that originally introduced multi-lets into the standard.