make `as Location` tests chain in a consistent way
Fixes #614.

This ensures that the `as Location` tests chain in a consistent way.
The parent directory of the dhall-lang repository is chosen because a
common use case is that test harness that runs these tests will probably
be in the host language repository and dhall-lang is pulled in as a
submodule.  (For an example of this, see spec_test.go in
philandstuff/dhall-golang)