Test import alternative with type error (#662)
This tests that the import alternative operator '?' successfully
recovers from type errors, i.e. if BAD=`0 0` then `env:BAD ? 0` should
evaluate to `0` and _not_ fail with a type error. The relevant section
of the standard covers this case:

Resolution of expressions might not be always successful: [..] imports
might not resolve in cases like:
- an environment variable is not defined
- [..]
- typecheck error
By using the ? operator, expressions are alternatively resolved [..]