Add Integer/toSigned
This new builtin splits an Integer apart into a Natural, and an encoding
of its sign.

There are a few points of motivation:

1. We have Integer/show, but this could be decomposed into
   Integer/toSigned and Natural/show.

2. We can represent exact rational numbers via the type
   { numerator : Natural, denominator : Natural }. However, constructing
   values of this type is difficult. One familiar way to construct such
   a value would be to use scientific notation. For example, 0.15625
   could be written as `e 15625 -5`.

3. https://github.com/dhall-lang/dhall-lang/issues/508 wants a way to
   show Integers signed, but only showing the sign when it's
   negative (that is, no leading +). This cannot be done with Integers
   currently, forcing this work to invent its own signed integer type.

This does open the door to exposing the total ordering of Integers (as
we now expose enough to determine the total ordering of Naturals). It
also gives you *some* form of Integer arithmetic, but you can't
reconstruct an Integer from it's signed type.