Don't URL-decode path segments (#704)
Currently, the standard requires implementations to URL-decode unquoted
path segments.  I think this is a mistake.

In particular, the URLs http://example.com/foo:bar and
http://example.com/foo%3Abar are potentially different URLs, even though
`:` is URL-encoded as `%3A`.  They are different because `:` is a
reserved character, which means that servers may interpret the `:` as a
special delimiter rather than the literal colon character.

An example of this kind of processing is the "matrix parameters" in
http://example.com/foo;bar=baz/quux -- matrix parameters are paremeters
on a path component, similar to how a query string contains parameters
on the whole URL.  Matrix parameters use the reserved characters `;` and
`=` as delimiter syntax.  If you want to use these characters within a
matrix parameter name or value, you need to percent-encode them.  This
demonstrates why `;` and `%3B` may not be equivalent in path segments.

By requiring implementations to URL-decode path segments, we
unintentionally fold distinct URLs into the same representation.

This commit fixes that by changing the internal representation to be the
original (encoded) text, and by requiring *quoted* path segments to be
URL-*encoded* to match.  This changes the binary representation for
URIs, but the binary representation for URIs is rarely used because it
cannot appear in a resolved and normalized expression.

Reserved characters are defined here: https://tools.ietf.org/html/rfc3986#section-2.2

Matrix URIs: https://www.w3.org/DesignIssues/MatrixURIs.html

Related to #581 and dhall-lang/dhall-haskell#1109.