change how we run `normalization` tests
I've had a few normalization bugs in dhall-golang which were not caught
by the normalization tests.  I think that this is because I am running
them according to the README: I normalize both sides, then compare the
results.  But if normalization behaviour is wrong, it can normalize both
`A` and `B` in the incorrect way, and the test will still pass.

By changing the tests to only normalize `A` (and not `B`), the tests are
much more effective - because the `B` side now specifies the expected
normal form, not merely a different expression with the same normal form
as `A`.