Various test fixes (#716)
* fix tests to match beta-normalization.md

As far as I can work out, these test missed some simplifications
required by the standard.

* remove duplicate test

ironically, duplicateFields.dhall is a near-duplicate of
RecordLitDuplicateFields.dhall

* change how we run `normalization` tests

I've had a few normalization bugs in dhall-golang which were not caught
by the normalization tests.  I think that this is because I am running
them according to the README: I normalize both sides, then compare the
results.  But if normalization behaviour is wrong, it can normalize both
`A` and `B` in the incorrect way, and the test will still pass.

By changing the tests to only normalize `A` (and not `B`), the tests are
much more effective - because the `B` side now specifies the expected
normal form, not merely a different expression with the same normal form
as `A`.

* make test match standard

This test failed for me, because my code was outputting something with
`λ(x : Natural) → x + 1` but the test expected `λ(n : Natural) → n +
1`.  Now, I can see how this is arguable as to whether my test should
have failed.  But note:

 - tests/README.md just asks if the results "match"; it does not say
   "are judgmentally equal"
 - similarly, it says to "normalize" the expressions, but doesn't say
   "alpha normalize"
 - standard/beta-normalization.md specifies that Natural/build passes
   `λ(x : Natural) → x + 1` to its argument

so I think a reasonable implementation of these tests can fail.

This fixes NaturalBuildImplementationB.dhall to match the standard.

* consistent test naming convention

* move parser/success/ifThenElse to unit/

this is a unit test, we should mark it as such

* add binary-decode tests for assert and ≡