Standardize support for record override
Fixes https://github.com/dhall-lang/dhall-lang/issues/382

This adds a new record instantiation operator of the form
`T::r` which is syntactic sugar for `(T.default ⫽ r) : T.Type)

The motivation for this change is to handle configuration formats
with a large number of defaultable fields (especially if the field
is `Optional` and the default is `None U`) so that users can omit any
all default-valued fields when creating the record.

```
let Example =
      { Type = { foo : Text, bar : Optional Natural }
      , default = { bar = None Natural }
      }

    -- No need to specify `bar` since it will default to `None Natural`
in  Example::{ foo = "ABC" }
```

This syntax more closely resembles the common JSON/YAML idiom of
omitting default-valued fields.