Simplify README by linking to integrations wiki page (#776)
Most of the README is superseded by the integrations page which contains
up-to-date information on integrating Dhall into various languages and
formats.  This change removes those stale sections in favor of the Wiki
page.

This condenses the README substantially to the point where the table of
contents is no longer necessary, either.