Import wiki contents to a sphinx-doc website
This change imports the wiki contents and use sphinx to render
an HTML website. The main features provided by sphinx I could not easily
implements in other website generators are:

* Table of content and navigation
* Linkable section heading with anchor
* Search

However sphinx (docutils) is best use with ReStructuredText as it is
arguably better suited for technical documentation. If using another markup
format is an option, then I vote for rST (and keep sphinx) or go with asciidoc.