Clarify process for merging changes after edits (#876)
Motivated by this discussion: https://github.com/dhall-lang/dhall-lang/pull/875#issuecomment-570818061

We've had a few cases recently where new contributors are unsure what the
process is if they make changes in response to review.  In particular, they
aren't sure if the clock resets after any changes or if new changes need to
be re-approved.

This proposed that the 3-day/7-day waiting periods are relative to the
original submission, and any edits after that only warrant a 24 hour
waiting period for subsequent edits (and changes don't require re-approval).