Add myself as a user with root privileges on `dhall-lang.org`
This is the first step in making it easier for other people to
manage this machine.  The full set of steps I plan after this are:

* Switch from using NixOps to using NixOS

* Install a service to automatically have `dhall-lang.org` track
  `master`

If you would like to also maintain this machine you can create a pull
request like this one to add yourself as an administrative user.