Add test for natural literals with leading zeros
After #859 added hexadecimal literals, the most obvious and simple way I
could have implemented this in Go would have been
`strconv.ParseInt(text, 0, 0)` [1].  This standard library function
automatically detects the number base based on prefix.

This would handle the typical decimal and hexadecimal cases well, but it
would not handle the case of a decimal number with a leading 0.  Dhall
treats this as decimal, but ParseInt treats this as octal.

Because this is such a common convention (C uses it and has had great
influence in this regard), I suspect many standard libraries will have
tempting utility functions, and that therefore it's wise to have a test
that explicitly checks that a number such as 042 parses as decimal.

(Another option would be to standardise octal literals instead but I am
neutral leaning slightly negative on this).

[1]: https://golang.org/pkg/strconv/#ParseInt