Fix typing rule for merging `Optional`s
The current rule for typechecking merge on `Optional`s normalizes the argument in order to know if it's a `None` or a `Some`, and uses that to typecheck the resulting `merge` expression.
Not only is that unnecessary work, I think it's actually wrong. The correct behavior is not to construct a value of type `< None | Some : A >` manually, but to assume one is present in the context.
For example, I believe the current rule fails to typecheck the following:
```dhall
λ(x : Optional Bool) → merge { None = 0, Some = λ(_ : Bool) → 1 } x
```
because it will try and fail to normalize `x` to either `None` or `Some`.