Don't parse single-quote strings too greedily
@Duncaen opened philandstuff/dhall-golang#1 with what I think is an ABNF
bug.  Broadly, in the example he gave (which I have used as a parser
test here), dhall-golang was parsing the single-quoted string too
greedily - taking the whole input to be a single-quoted string, instead
of being two single-quoted strings separated by a text-append operator.

He also suggested the fix: a reordering of the ABNF rules to treat the
string `''` as the end of a single-quote string, unless it is the start
of the special sequences `'''` or `''${`.  I think this fix is correct.

I suspect the reason this bug affected dhall-golang but not
dhall-haskell is that dhall-golang uses a PEG parser, and so rigorously
applies alternatives in strict order?

Co-authored-by: Duncan Overbruck <mail@duncano.de>