Add support for `with` keyword
Fixes https://github.com/dhall-lang/dhall-lang/issues/114

Fixes https://github.com/dhall-lang/dhall-lang/issues/340

Related to https://github.com/dhall-lang/dhall-lang/issues/754

This adds a new keyword which is syntactic sugar for using `//` to
perform a deep override of the record.  This comes in handy when
working with very large records (such as those found in
`dhall-kubernetes`).  This similar in spirit to Kustomize, albeit not
as powerful and customizable as Kustomize.

The goal is that the user can write the following code to update an
existing default configuration:

```dhall
someDeployment with {
, metadata.labels =
      someDeployment.metadata.labels
    # toMap { `scheduler.alpha.kubernetes.io/critical-pod` = "" }
}
```

You can also see this idiom outside of the Dhall/Kuberentes ecosystem.
For example, many Nix users manually apply deeply nested updates by
hand, like this example from the Nixpkgs manual:

```nix
  {
    haskell = super.haskell // {
      packages = super.haskell.packages // {
        ghc784 = super.haskell.packages.ghc784.override {
          overrides = self: super: {
            ghc-events = self.callPackage ./ghc-events-0.4.3.0.nix {};
          };
        };
      };
    };
  };
```

... where the Dhall equivalent would be:

```dhall
super with
  { haskell.packages.ghc784 = super.haskell.packages.ghc784.override {
      …
    }
  }
```

There is one additional change that I plan to make that builds on top of
this which is to would like to have the record completion syntax desugar
to use `with` instead of `//`, like this:

```dhall
T::r = (T.default with r) : T.Type
```

... which would allow for deep overrides for the record completion
operator, too.  However, I chose to defer that to a separate change to
keep this change as minimal as possible.