Add test that we ignore poisoned cache entries
The standard requires implementations to check the hash of files fetched
from the cache:

    Similarly, an implementation MUST follow these steps when importing an
    expression protected by a semantic integrity check:

    * Check if there is a Dhall expression stored at either
      `"${XDG_CACHE_HOME}/dhall/1220${base16Hash}"` or
      `"${HOME}/.cache/dhall/1220${base16Hash}"`
    * If the file exists and is readable, verify the file's byte contents match the
      hash and then decode the expression from the bytes using the `decode` judgment
      instead of importing the expression
    * Otherwise, import the expression as normal

The feature is also noted in the import judgments.  However, this feature
is easy to miss (and indeed dhall-golang did miss it).  This adds a test
to ensure we ignore such poisoned cache entries.

This feature will become even more important if the [dhall proxy][1]
becomes a thing, and we use remote caches stored on untrusted servers.

[1]: https://discourse.dhall-lang.org/t/rfc-proxy-dhall-lang-org/144/2