[tests] Remove dependency on csrng.net (#969)
The tests currently depend on csrng.net, which has a rate limit of one
request per second.  This is slow enough that my tests hit the limit
consistently.

Returning a random string from a URL should be doable in nginx config,
so this commit adds a module and a location block to return a random
string from https://test.dhall-lang.org/random-string for testing
purposes.