Create `store.dhall-lang.org` (#990)
For more details, see:

https://discourse.dhall-lang.org/t/rfc-proxy-dhall-lang-org/144/3?u=gabriel439

The short summary is that we can declaratively add Dhall packages to
cache and make that cache available from `store.dhall-lang.org`.

The path of a cached item is the item's multihash.  So, for example,
the address of version 13.0.0 of the Prelude would be:

https://store.dhall-lang.org/12204aa8581954f7734d09b7b21fddbf5d8df901a44b54b4ef26ea71db92de0b1a12

Interpreters can then optionally avail themselves of this shared cache,
either automatically or with explicit user consent, in order to speed up
package imports.