Ensure `keyword` rule only matches keywords
The `keyword` rule is provided for PEG (and similar) parsers to exclude
keywords from places like `simple-label` - and the `simple-label` rule
has an example above for how to do this.

However, `keyword` matches `∀`, which means that a naive use of the
`keyword` rule in `simple-label` would permit simple labels such as
`∀bcde`, which obviously should not be permitted.

This commit splits up the `forall` rule so that `keyword` can match the
forall keyword but not the forall symbol.