Optimize desugaring of `with` expressions
The motivation for this change is:

https://github.com/dhall-lang/dhall-haskell/issues/1960

… where a deeply nested and chained `with` expression caused a blow-up
in the size of the desugared syntax tree.

This alternative desugaring was actually an idea we contemplated in the
original change that standardized `with`:

https://github.com/dhall-lang/dhall-lang/pull/923

… but at the time we elected to go with the simpler desugaring, reasoning
that users could work around performance issues with explicit `let`
binding:

https://github.com/dhall-lang/dhall-lang/pull/923#issuecomment-593116404

However, for chained `with` expressions that work-around is not
ergonomic, as illustrated in
https://github.com/dhall-lang/dhall-haskell/issues/1960.  Instead,
this adds language support for the `let` work-around.

The change to the `WithMultiple` test demonstrates that this leads to
a more compact desugared expression.  Before this change the
expression `{ a.b = 1, c.d = 2 } with a.b = 3 with c.e = 4` would
desugar to:

```dhall
  { a.b = 1, c.d = 2 }
⫽ { a = { a.b = 1, c.d = 2 }.a ⫽ { b = 3 } }
⫽ { c =
        ({ a.b = 1, c.d = 2 } ⫽ { a = { a.b = 1, c.d = 2 }.a ⫽ { b = 3 } }).c
      ⫽ { e = 4 }
  }
```

… and now the same expression desugars to:

```dhall
let _ = let _ = { a.b = 1, c.d = 2 } in _ ⫽ { a = _.a ⫽ { b = 3 } }

in  _ ⫽ { c = _.c ⫽ { e = 4 } }
```