Enable `with` optimizations (#1052)
The motivation for this change is:

https://github.com/dhall-lang/dhall-haskell/issues/1960

This change to the standard gives implementations greater freedom
in how they desugar a `with` expression, in order to permit more
efficient implementations.  The standard now also suggests a more
efficient implementation to help implementation authors.

This is technically a breaking change to the binary encoding of
a non-normalized `with` expression, but most semantic integrity checks
are unaffected by this change.  The exception are those where
an imported expression contains a `with` expression with an
abstract left-hand side (see the `WithDesugar` test for an example of
code that would have a new integrity check after this change)