Literate Haskell Semantics - Syntax
Related to: https://github.com/dhall-lang/dhall-lang/issues/959

This is the first step towards an executable reference implementation by
creating a Haskell project which contains the Haskell type for the AST.

The Haskell project is built by CI, in order to catch any errors.

Note that the long-term goal is to standardize a normalization-by-evaluation
approach for the semantics in order to make it easier for
implementers to build efficient implementations.  However, in the short-term
I'm translating the semantics as literally as possible to Haskell
code and then I'll do a separate pass later on to switch to
normalization-by-evaluation.