Improve `Prelude.JSON.render` output
Now that we have language support for `Text/replace` we don't need
to use `Text/show` any longer.  In particular, one of the limitations
of `Text/show` was that we were using it both to render valid Dhall
values and also to render valid JSON strings, which constrained it
to produce weird output for `$` (`\u0024`).  However, now that we
have `Text/replace` we can use a JSON-specific escaping strategy.

This also implies that we can make two further improvements (not
included in this pull request):

* We could standardize the behavior of `Text/show` in terms of
  `Text/replace`

  … or even remove `Text/show` as a built-in after deprecation.

* We can now improve the behavior of `Text/show` to escape `$` as
  `\$` instead of `\u0024`

  … now that it's no longer used for escaping JSON strings.