Fix Hydra builds (#1117)
Hydra started to get stuck after being upgraded in #1114.  Specifically,
builds would get stuck on the "Receiving outputs" step.

The reason this was happening was because of how we were instructing
Hydra to build things on the local machine.  The correct way to do this
is to specify `localhost`, because Hydra has special logic to handle
that.  If you specify the current machine by any other hostname then
Hydra will treat it as if it is a remote machine and then get stuck
attempting to take the same lock twice (once for its own store and once
for the "remote" machine's store, which happens to be the same store).

Also, now that we don't `ssh` into our own machine we don't need to generate a
key pair for the `hydra-queue-runner` user, and we also don't need to
install the public key for our host as a known host.