Rotate Discourse SMTP password (#1138)
In the course of debugging the issues that led to
https://github.com/dhall-lang/dhall-lang/pull/1137 I accidentally
deleted `nixops/discourseSmtpPassword` on dhall-lang.org using a
`git clean`, so I had to create a new password.  This pull request
updates the password that the SMTP mailserver expects for the
Discourse mail account.

I also removed `nixops/discourseSmtpPassword` from `.gitignore`, to
make it harder for me to make this mistake in the future (since
it will now show up in the list of untracked files which I check
before doing a `git clean`).