Add reference implementation of parser
The main motivation for this change is to enable testing the reference
implementation before proceeding any further, to make it easier for
people to review changes to the code.

This change adds a reference implementation of the parser using an
LL parsing package (`megaparsec`) and adds a test suite to verify that
the parser passes all of the parsing success acceptance tests.  I have
not yet checked any of the parsing failure tests.

This also begins adding an executable for ease of testing.  Currently
all the executable does is print the binary encoding of an expression.