Change `?` to only fall back on absent imports (#1181)
Fixes https://github.com/dhall-lang/dhall-lang/issues/1113

This changes the behavior of the `?` operator to only fall back to
the second import if it cannot be retrieved (either from the
cache or the specified location).  In particular, it will not
recover from type errors, parse errors, cyclic imports, or hash
mismatches.