Update Dhall (#1228)
The main motivation for this change is so that Mac users can run
`./scripts/{generate-test-files,lint-prelude}.sh`.  The reason this
fixes things is that the latest version of `dhall-haskell` supports
Big Sur because it upgrades its Nixpkgs pin.

We have to use an older version of Dhall for dhall-lang.org, though,
until Nixpkgs supports GHCJS again.  However, that shouldn't be too
big of a deal since the website examples are stable and don't require
newer language features.