dhall-lang.org: Fix continuous deployment
... to fix the issue reported in:

https://github.com/dhall-lang/dhall-lang/pull/1296#issuecomment-1227679350

What happened is that due to this bug in `nixos-mailserver`:

https://gitlab.com/simple-nixos-mailserver/nixos-mailserver/-/issues/232

... the `switch-to-configuration` step of `self-deploy` would hang
permanently, preventing the system from picking up new changes.

However, fixing that required updating the version of `nixos-mailserver`
to the 22.05 branch, which in turn required updating the version of
`nixpkgs` to 22.05, too.

That in turn broke the `codemirror` build, so I switched to downloading
a prebuilt `codemirror` archive instead of building `codemirror` using
`node2nix`.

Also, we now no longer need to use `pkgs.hydra-unstable` now that we've
upgraded to a newer version of Nixpkgs.