Prelude: Split `file` constructor of `DirectoryTree.Make` (#1398)
* Splitted `file` constructor of DirectoryTree

This commit replaces the `file` construction function of Directory.Make
with `text-file` and `binary-file`. The former is what `file` used to
be, i.e. it constructs files with `Text` content.
The new `binary-file` constructor is suitable for creating files with
`Bytes` content.

* DirectoryTree: Do not remove `file` constructor

Only add `binary-file` so we stay backwards compatible.

* Updated import hashes

* Added an alternative code path for non-Nix users to lint-prelude script